+++ src/main/java/net/andrecarbajal/naviMusic/commands/ICommand.java
@@ -1,0 +1,13 @@
+package net.andrecarbajal.naviMusic.commands;
+
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
+import net.dv8tion.jda.api.interactions.commands.build.OptionData;
+
+import java.util.List;
+
+public interface ICommand {
+    String getName();
+    String getDescription();
+    List<OptionData> getOptions();
+    void execute(SlashCommandInteractionEvent event);
+}