+++ src/main/java/net/andrecarbajal/naviMusic/lavaplayer/PlayerManager.java
@@ -1,0 +1,50 @@
+package net.andrecarbajal.naviMusic.lavaplayer;
+
+import com.sedmelluq.discord.lavaplayer.player.AudioPlayerManager;
+import com.sedmelluq.discord.lavaplayer.player.DefaultAudioPlayerManager;
+import com.sedmelluq.discord.lavaplayer.source.AudioSourceManagers;
+import net.andrecarbajal.naviMusic.lavaplayer.handler.PlaylistLoadResultHandler;
+import net.andrecarbajal.naviMusic.lavaplayer.handler.TrackLoadResultHandler;
+import net.dv8tion.jda.api.entities.Guild;
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class PlayerManager {
+    private static PlayerManager INSTANCE;
+    private final Map<Long, GuildMusicManager> guildMusicManagers = new HashMap<>();
+    private final AudioPlayerManager audioPlayerManager = new DefaultAudioPlayerManager();
+
+    private PlayerManager() {
+        AudioSourceManagers.registerRemoteSources(audioPlayerManager);
+        AudioSourceManagers.registerLocalSource(audioPlayerManager);
+    }
+
+    public static PlayerManager get() {
+        if(INSTANCE == null) {
+            INSTANCE = new PlayerManager();
+        }
+        return INSTANCE;
+    }
+
+    public GuildMusicManager getGuildMusicManager(Guild guild) {
+        return guildMusicManagers.computeIfAbsent(guild.getIdLong(), (guildId) -> {
+            GuildMusicManager musicManager = new GuildMusicManager(audioPlayerManager, guild);
+
+            guild.getAudioManager().setSendingHandler(musicManager.getAudioForwarder());
+
+            return musicManager;
+        });
+    }
+
+    public void play(Guild guild, String trackURL, SlashCommandInteractionEvent event) {
+        GuildMusicManager guildMusicManager = getGuildMusicManager(guild);
+        audioPlayerManager.loadItemOrdered(guildMusicManager, trackURL, new TrackLoadResultHandler(guildMusicManager, event));
+    }
+
+    public void loadAndPlayPlaylist(Guild guild, String trackUrl, SlashCommandInteractionEvent event) {
+        GuildMusicManager musicManager = getGuildMusicManager(guild);
+        audioPlayerManager.loadItemOrdered(guild, trackUrl, new PlaylistLoadResultHandler(musicManager, event, trackUrl));
+    }
+}