+++ RELEASE_NOTES.md
@@ -1,4 +1,3 @@
 # Release Notes
 
-## Bug Fixes
-- when searching for a song, it is now done through Youtube Music and no longer through youtube.
+- Package name changed from `anvian` to `andrecarbajal`.